+++ decidim-admin/app/events/decidim/attachment_created_event.rb
@@ -1,42 +1,8 @@
 # frozen-string_literal: true
 
 module Decidim
-  class AttachmentCreatedEvent < Decidim::Events::BaseEvent
-    include Decidim::Events::EmailEvent
-    include Decidim::Events::NotificationEvent
-
-    def email_subject
-      I18n.t(
-        "decidim.events.attachments.attachment_created.email_subject",
-        resource_title: resource_title,
-        resource_path: resource_path
-      )
-    end
-
-    def email_intro
-      I18n.t(
-        "decidim.events.attachments.attachment_created.email_intro",
-        resource_title: resource_title,
-        resource_path: resource_path
-      )
-    end
-
-    def email_outro
-      I18n.t(
-        "decidim.events.attachments.attachment_created.email_outro",
-        resource_title: resource_title,
-        resource_path: resource_path
-      )
-    end
-
-    def notification_title
-      I18n.t(
-        "decidim.events.attachments.attachment_created.notification_title",
-        resource_title: resource_title,
-        resource_path: resource_path,
-        attached_to_url: attached_to_url
-      ).html_safe
-    end
+  class AttachmentCreatedEvent < Decidim::Events::SimpleEvent
+    i18n_attributes :attached_to_url
 
     def resource_path
       @resource.url