+++ decidim_app-design/app/helpers/application_helper.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 module ApplicationHelper
-  def partial(name)
-    render partial: "#{params[:layout]}/partials/#{name}"
+  def partial(name, opts = {})
+    render partial: "#{params[:layout]}/partials/#{name}", locals: opts
   end
 
   def page_path(name)